Move symbol propagation into separate file
